+++ src/ops/mod.rs
@@ -1,8 +1,7 @@
 pub use self::blit::blit;
 pub use self::clear::clear;
 pub use self::draw::draw;
-pub use self::read::{read_attachment, read_from_default_fb};
-pub use self::read::{read_attachment_to_pb, read_from_default_fb_to_pb};
+pub use self::read::{read, read_if_supported, Source, Destination};
 
 mod blit;
 mod clear;