+++ src/ops/read.rs
@@ -1,142 +1,153 @@
 use std::ptr;
 
-use pixel_buffer::{self, PixelBuffer};
+use pixel_buffer::PixelBuffer;
 use texture::ClientFormat;
-
-use context::Context;
-use ContextExt;
+use texture::PixelValue;
 
 use fbo;
-use texture;
 
 use GlObject;
-use libc;
+use Rect;
 use context::CommandContext;
 use gl;
 
-pub fn read_attachment<P, T>(attachment: &fbo::Attachment, dimensions: (u32, u32),
-                             context: &Context) -> T          // TODO: remove Clone for P
-                             where P: texture::PixelValue + Clone + Send,
-                             T: texture::Texture2dDataSink<Data = P>
-{
-    let mut ctxt = context.make_current();
-
-    let (fbo, atch) = context.get_framebuffer_objects()
-                             .get_framebuffer_for_reading(attachment, &mut ctxt);
-
-    read_impl(fbo, atch, dimensions, None, &mut ctxt).unwrap()
+/// A source for reading pixels.
+pub enum Source<'a> {
+    // TODO: remove the second parameter
+    Attachment(&'a fbo::Attachment, &'a fbo::FramebuffersContainer),
+    // TODO: use a Rust enum
+    DefaultFramebuffer(gl::types::GLenum),
 }
 
-/// Panics if the pixel buffer is not big enough.
-pub fn read_attachment_to_pb<P, T>(attachment: &fbo::Attachment, dimensions: (u32, u32),
-                                   dest: &mut PixelBuffer<T>, context: &Context)          // TODO: remove Clone for P
-                                   where P: texture::PixelValue + Clone + Send,
-                                   T: texture::Texture2dDataSink<Data = P>
-{
-    let mut ctxt = context.make_current();
-
-    let (fbo, atch) = context.get_framebuffer_objects()
-                             .get_framebuffer_for_reading(attachment, &mut ctxt);
-
-    read_impl(fbo, atch, dimensions, Some(dest), &mut ctxt);
-}
-
-pub fn read_from_default_fb<P, T>(attachment: gl::types::GLenum, context: &Context) -> T          // TODO: remove Clone for P
-                                  where P: texture::PixelValue + Clone + Send,
-                                  T: texture::Texture2dDataSink<Data = P>
-{
-    let mut ctxt = context.make_current();
-
-    let (w, h) = context.get_framebuffer_dimensions();
-    let (w, h) = (w as u32, h as u32);      // TODO: remove this conversion
-    read_impl(0, attachment, (w, h), None, &mut ctxt).unwrap()
-}
-
-/// Panics if the pixel buffer is not big enough.
-pub fn read_from_default_fb_to_pb<P, T>(attachment: gl::types::GLenum,
-                                        dest: &mut PixelBuffer<T>, context: &Context)          // TODO: remove Clone for P
-                                        where P: texture::PixelValue + Clone + Send,
-                                        T: texture::Texture2dDataSink<Data = P>
-{
-    let mut ctxt = context.make_current();
-    let (w, h) = context.get_framebuffer_dimensions();
-    read_impl(0, attachment, (w, h), Some(dest), &mut ctxt);
-}
-
-fn read_impl<P, T>(fbo: gl::types::GLuint, readbuffer: gl::types::GLenum,
-                   dimensions: (u32, u32), target: Option<&mut PixelBuffer<T>>,
-                   mut ctxt: &mut CommandContext) -> Option<T>          // TODO: remove Clone for P
-                   where P: texture::PixelValue + Clone + Send,
-                   T: texture::Texture2dDataSink<Data = P>
-{
-    use std::mem;
-
-    let pixels_count = dimensions.0 * dimensions.1;
-
-    let chosen_format = <T as texture::Texture2dDataSink>::get_preferred_formats()[0];
-    let pixels_size = chosen_format.get_size();
-    let (format, gltype) = client_format_to_gl_enum(&chosen_format);
-
-    let total_data_size = pixels_count as usize * pixels_size;
-
-    let pixel_buffer = target.as_ref().map(|buf| buf.get_id()).unwrap_or(0);
-
-    if let Some(pixel_buffer) = target {
-        assert!(pixel_buffer.get_size() >= total_data_size);
-        pixel_buffer::store_infos(pixel_buffer, dimensions, chosen_format);
+// TODO: re-enable when the second parameter is no longer needed
+/*impl<'a> From<&'a fbo::Attachment> for Source<'a> {
+    fn from(a: &'a fbo::Attachment) -> Source<'a> {
+        Source::Attachment(a)
     }
+}*/
 
-    let data = unsafe {
+/// A destination for reading pixels.
+pub enum Destination<'a, P> where P: PixelValue {
+    Memory(&'a mut Vec<P>),
+    PixelBuffer(&'a PixelBuffer<P>),
+    // TODO: texture with glCopyTexSubImage2D
+}
+
+impl<'a, P> From<&'a mut Vec<P>> for Destination<'a, P> where P: PixelValue {
+    fn from(mem: &'a mut Vec<P>) -> Destination<'a, P> {
+        Destination::Memory(mem)
+    }
+}
+
+impl<'a, P> From<&'a PixelBuffer<P>> for Destination<'a, P> where P: PixelValue {
+    fn from(pb: &'a PixelBuffer<P>) -> Destination<'a, P> {
+        Destination::PixelBuffer(pb)
+    }
+}
+
+/// Reads pixels from the source into the destination.
+///
+/// Panicks if the destination is not large enough.
+///
+/// The `(u8, u8, u8, u8)` format is guaranteed to be supported.
+pub fn read<'a, S, D>(mut ctxt: &mut CommandContext, source: S, rect: &Rect, dest: D)
+                      where S: Into<Source<'a>>, D: Into<Destination<'a, (u8, u8, u8, u8)>>
+{
+    match read_if_supported(ctxt, source, rect, dest) {
+        Ok(_) => (),
+        Err(_) => unreachable!(),
+    }
+}
+
+/// Reads pixels from the source into the destination.
+///
+/// Panicks if the destination is not large enough.
+pub fn read_if_supported<'a, S, D, T>(mut ctxt: &mut CommandContext, source: S, rect: &Rect,
+                                      dest: D) -> Result<(), ()>
+                                      where S: Into<Source<'a>>, D: Into<Destination<'a, T>>,
+                                            T: PixelValue
+{
+    let source = source.into();
+    let dest = dest.into();
+
+    let pixels_to_read = rect.width * rect.height;
+
+    let (fbo, read_buffer) = match source {
+        Source::Attachment(attachment, framebuffer_objects) => {
+            framebuffer_objects.get_framebuffer_for_reading(attachment, &mut ctxt)
+        },
+        Source::DefaultFramebuffer(read_buffer) => {
+            (0, read_buffer)
+        },
+    };
+
+    // FIXME: check if format is suppoed by ReadPixels
+
+    let (format, gltype) = client_format_to_gl_enum(&<T as PixelValue>::get_format());
+
+    unsafe {
         // binding framebuffer
+        // FIXME: GLES2 only supports color attachment 0
         fbo::bind_framebuffer(&mut ctxt, fbo, false, true);
 
         // adjusting glReadBuffer
-        ctxt.gl.ReadBuffer(readbuffer);
-
-        // adjusting data alignement
-        if ctxt.state.pixel_store_pack_alignment != 1 {
-            ctxt.state.pixel_store_pack_alignment = 1;
-            ctxt.gl.PixelStorei(gl::PACK_ALIGNMENT, 1);
-        }
-
-        // binding buffer
-        if ctxt.state.pixel_pack_buffer_binding != pixel_buffer {
-            ctxt.gl.BindBuffer(gl::PIXEL_PACK_BUFFER, pixel_buffer);
-            ctxt.state.pixel_pack_buffer_binding = pixel_buffer;
-        }
+        // FIXME: handle this in `fbo`
+        ctxt.gl.ReadBuffer(read_buffer);
 
         // reading
-        if pixel_buffer == 0 {
-            let data_size = pixels_count as usize * pixels_size / mem::size_of::<P>();
-            let mut data: Vec<P> = Vec::with_capacity(data_size);
-            ctxt.gl.ReadPixels(0, 0, dimensions.0 as gl::types::GLint,
-                               dimensions.1 as gl::types::GLint, format, gltype,
-                               data.as_mut_ptr() as *mut libc::c_void);
-            data.set_len(data_size);
-            Some(data)
+        match dest {
+            Destination::Memory(dest) => {
+                let mut buf = Vec::with_capacity(pixels_to_read as usize);
 
-        } else {
-            ctxt.gl.ReadPixels(0, 0, dimensions.0 as gl::types::GLint,
-                               dimensions.1 as gl::types::GLint, format, gltype,
-                               ptr::null_mut());
-            None
+                // FIXME: correct function call
+                if ctxt.state.pixel_pack_buffer_binding != 0 {
+                    ctxt.gl.BindBuffer(gl::PIXEL_PACK_BUFFER, 0);
+                    ctxt.state.pixel_pack_buffer_binding = 0;
+                }
+
+                // adjusting data alignement
+                let ptr = buf.as_mut_ptr() as *mut D;
+                let ptr = ptr as usize;
+                if (ptr % 8) == 0 {
+                } else if (ptr % 4) == 0 && ctxt.state.pixel_store_pack_alignment != 4 {
+                    ctxt.state.pixel_store_pack_alignment = 4;
+                    ctxt.gl.PixelStorei(gl::PACK_ALIGNMENT, 4);
+                } else if (ptr % 2) == 0 && ctxt.state.pixel_store_pack_alignment > 2 {
+                    ctxt.state.pixel_store_pack_alignment = 2;
+                    ctxt.gl.PixelStorei(gl::PACK_ALIGNMENT, 2);
+                } else if ctxt.state.pixel_store_pack_alignment != 1 {
+                    ctxt.state.pixel_store_pack_alignment = 1;
+                    ctxt.gl.PixelStorei(gl::PACK_ALIGNMENT, 1);
+                }
+
+                ctxt.gl.ReadPixels(rect.left as gl::types::GLint, rect.bottom as gl::types::GLint,
+                                   rect.width as gl::types::GLsizei,
+                                   rect.height as gl::types::GLsizei, format, gltype,
+                                   buf.as_mut_ptr() as *mut _);
+                buf.set_len(pixels_to_read as usize);
+
+                *dest = buf;
+            },
+
+            Destination::PixelBuffer(pb) => {
+                assert!(pb.len() >= pixels_to_read as usize);
+
+                let pb = pb.get_id();
+                // FIXME: correct function call
+                if ctxt.state.pixel_pack_buffer_binding != pb {
+                    ctxt.gl.BindBuffer(gl::PIXEL_PACK_BUFFER, pb);
+                    ctxt.state.pixel_pack_buffer_binding = pb;
+                }
+
+                ctxt.gl.ReadPixels(rect.left as gl::types::GLint, rect.bottom as gl::types::GLint,
+                                   rect.width as gl::types::GLsizei,
+                                   rect.height as gl::types::GLsizei, format, gltype,
+                                   ptr::null_mut());
+            }
         }
     };
 
-    if let Some(data) = data {
-        let data = texture::RawImage2d {
-            data: ::std::borrow::Cow::Owned(data),
-            width: dimensions.0 as u32,
-            height: dimensions.1 as u32,
-            format: chosen_format,
-        };
-
-        Some(texture::Texture2dDataSink::from_raw(data))
-
-    } else {
-
-        None
-    }
+    Ok(())
 }
 
 fn client_format_to_gl_enum(format: &ClientFormat) -> (gl::types::GLenum, gl::types::GLenum) {