+++ tests/texture_write.rs
@@ -16,11 +16,11 @@ fn texture_2d_write() {
     texture.write(glium::Rect { bottom: 1, left: 1, width: 1, height: 1 },
                   vec![vec![(128u8, 64u8, 2u8)]]);
 
-    let read_back: Vec<Vec<(u8, u8, u8)>> = texture.read();
-    assert_eq!(read_back[0][0], (0, 1, 2));
-    assert_eq!(read_back[0][1], (4, 8, 16));
-    assert_eq!(read_back[1][0], (32, 64, 128));
-    assert_eq!(read_back[1][1], (128, 64, 2));
+    let read_back: Vec<Vec<(u8, u8, u8, u8)>> = texture.read();
+    assert_eq!(read_back[0][0], (0, 1, 2, 255));
+    assert_eq!(read_back[0][1], (4, 8, 16, 255));
+    assert_eq!(read_back[1][0], (32, 64, 128, 255));
+    assert_eq!(read_back[1][1], (128, 64, 2, 255));
 
     display.assert_no_error(None);
 }