+++ tests/uniforms.rs
@@ -44,9 +44,9 @@ fn uniforms_storage_single_value() {
     texture.as_surface().clear_color(0.0, 0.0, 0.0, 0.0);
     texture.as_surface().draw(&vb, &ib, &program, &uniforms, &Default::default()).unwrap();
 
-    let data: Vec<Vec<(u8, u8, u8)>> = texture.read();
-    assert_eq!(data[0][0], (255, 0, 0));
-    assert_eq!(data.last().unwrap().last().unwrap(), &(255, 0, 0));
+    let data: Vec<Vec<(u8, u8, u8, u8)>> = texture.read();
+    assert_eq!(data[0][0], (255, 0, 0, 128));
+    assert_eq!(data.last().unwrap().last().unwrap(), &(255, 0, 0, 128));
 
     display.assert_no_error(None);
 }
@@ -85,9 +85,9 @@ fn uniforms_storage_multiple_values() {
     texture.as_surface().clear_color(0.0, 0.0, 0.0, 0.0);
     texture.as_surface().draw(&vb, &ib, &program, &uniforms, &Default::default()).unwrap();
 
-    let data: Vec<Vec<(u8, u8, u8)>> = texture.read();
-    assert_eq!(data[0][0], (255, 0, 0));
-    assert_eq!(data.last().unwrap().last().unwrap(), &(255, 0, 0));
+    let data: Vec<Vec<(u8, u8, u8, u8)>> = texture.read();
+    assert_eq!(data[0][0], (255, 0, 0, 255));
+    assert_eq!(data.last().unwrap().last().unwrap(), &(255, 0, 0, 255));
 
     display.assert_no_error(None);
 }
@@ -126,9 +126,9 @@ fn uniforms_storage_ignore_inactive_uniforms() {
     texture.as_surface().clear_color(0.0, 0.0, 0.0, 0.0);
     texture.as_surface().draw(&vb, &ib, &program, &uniforms, &Default::default()).unwrap();
 
-    let data: Vec<Vec<(u8, u8, u8)>> = texture.read();
-    assert_eq!(data[0][0], (255, 0, 0));
-    assert_eq!(data.last().unwrap().last().unwrap(), &(255, 0, 0));
+    let data: Vec<Vec<(u8, u8, u8, u8)>> = texture.read();
+    assert_eq!(data[0][0], (255, 0, 0, 128));
+    assert_eq!(data.last().unwrap().last().unwrap(), &(255, 0, 0, 128));
 
     display.assert_no_error(None);
 }